search-experience-test-lighthouse/requirements.txt
crochet
lighthouse-python-plus
matplotlib
python-dateutil
python-dotenv
openai
psycopg2-binary
redis